# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nhyira's Atelier API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: portfolio
    description: Portfolio project operations
  - name: enquiries
    description: Enquiry operations
  - name: auth
    description: Admin authentication

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials

  /auth/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Admin logout
      responses:
        "200":
          description: Logged out

  /auth/me:
    get:
      operationId: getAdminMe
      tags: [auth]
      summary: Get current session
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Not authenticated

  /portfolio:
    get:
      operationId: listPortfolio
      tags: [portfolio]
      summary: List all portfolio items
      parameters:
        - name: published
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Portfolio list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioItem"

    post:
      operationId: createPortfolioItem
      tags: [portfolio]
      summary: Create a portfolio item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioItem"

  /portfolio/{id}:
    get:
      operationId: getPortfolioItem
      tags: [portfolio]
      summary: Get a portfolio item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Portfolio item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioItem"
        "404":
          description: Not found

    put:
      operationId: updatePortfolioItem
      tags: [portfolio]
      summary: Update a portfolio item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioItem"

    delete:
      operationId: deletePortfolioItem
      tags: [portfolio]
      summary: Delete a portfolio item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /enquiries:
    get:
      operationId: listEnquiries
      tags: [enquiries]
      summary: List all enquiries (admin)
      responses:
        "200":
          description: Enquiries list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enquiry"

    post:
      operationId: submitEnquiry
      tags: [enquiries]
      summary: Submit a public enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnquiryInput"
      responses:
        "201":
          description: Enquiry submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"

  /enquiries/{id}:
    get:
      operationId: getEnquiry
      tags: [enquiries]
      summary: Get a specific enquiry (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Enquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"

    patch:
      operationId: updateEnquiry
      tags: [enquiries]
      summary: Update enquiry status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnquiryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"

    delete:
      operationId: deleteEnquiry
      tags: [enquiries]
      summary: Delete an enquiry (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /stats:
    get:
      operationId: getStats
      tags: [enquiries]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string

    AuthResult:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean

    PortfolioItem:
      type: object
      required: [id, title, category, images, createdAt, published]
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        packageDetails:
          type: ["string", "null"]
        estimatedBudget:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean
        createdAt:
          type: string

    PortfolioInput:
      type: object
      required: [title, category, images]
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string
        packageDetails:
          type: string
        estimatedBudget:
          type: string
        images:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean

    PortfolioUpdate:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string
        packageDetails:
          type: string
        estimatedBudget:
          type: string
        images:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean

    Enquiry:
      type: object
      required: [id, name, email, service, message, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        service:
          type: string
        eventDate:
          type: ["string", "null"]
        message:
          type: string
        status:
          type: string
          enum: [new, read, replied, booked, archived]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    EnquiryInput:
      type: object
      required: [name, email, service, message]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        service:
          type: string
        eventDate:
          type: string
        message:
          type: string

    EnquiryUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [new, read, replied, booked, archived]
        notes:
          type: string

    DashboardStats:
      type: object
      required: [totalEnquiries, newEnquiries, bookedEnquiries, totalPortfolioItems]
      properties:
        totalEnquiries:
          type: integer
        newEnquiries:
          type: integer
        bookedEnquiries:
          type: integer
        totalPortfolioItems:
          type: integer
        enquiriesByService:
          type: array
          items:
            type: object
            required: [service, count]
            properties:
              service:
                type: string
              count:
                type: integer
        recentEnquiries:
          type: array
          items:
            $ref: "#/components/schemas/Enquiry"