# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nhyira's Atelier API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardStatsEnquiriesByServiceItem } from './dashboardStatsEnquiriesByServiceItem';
import type { Enquiry } from './enquiry';

export interface DashboardStats {
  totalEnquiries: number;
  newEnquiries: number;
  bookedEnquiries: number;
  totalPortfolioItems: number;
  enquiriesByService?: DashboardStatsEnquiriesByServiceItem[];
  recentEnquiries?: Enquiry[];
}