# lib/api-zod/src/generated/types/enquiry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nhyira's Atelier API
 * OpenAPI spec version: 0.1.0
 */
import type { EnquiryStatus } from './enquiryStatus';

export interface Enquiry {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  service: string;
  /** @nullable */
  eventDate?: string | null;
  message: string;
  status: EnquiryStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}