# lib/api-zod/src/generated/types/enquiryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nhyira's Atelier API
 * OpenAPI spec version: 0.1.0
 */
import type { EnquiryUpdateStatus } from './enquiryUpdateStatus';

export interface EnquiryUpdate {
  status?: EnquiryUpdateStatus;
  notes?: string;
}