# lib/api-zod/src/generated/types/portfolioInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nhyira's Atelier API
 * OpenAPI spec version: 0.1.0
 */

export interface PortfolioInput {
  title: string;
  category: string;
  description?: string;
  packageDetails?: string;
  estimatedBudget?: string;
  images: string[];
  tags?: string[];
  published?: boolean;
}