# lib/api-zod/src/generated/types/portfolioItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nhyira's Atelier API
 * OpenAPI spec version: 0.1.0
 */

export interface PortfolioItem {
  id: number;
  title: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  packageDetails?: string | null;
  /** @nullable */
  estimatedBudget?: string | null;
  images: string[];
  tags?: string[];
  published: boolean;
  createdAt: string;
}